refactor: Services to flatten interface and impl
Having only a single implementation for an interface is redundant,
flatten to the interfact and implementations to a single implementation
class with the original service interface's naming.

NO-TICKET
