+++ build.gradle
@@ -12,7 +12,7 @@ plugins {
 }
 
 group = "uk.nhs.hee.trainee.details"
-version = "0.6.0"
+version = "0.6.1"
 sourceCompatibility = "11"
 
 configurations {
