--- src/main/java/uk/nhs/hee/trainee/details/service/impl/PersonalDetailsServiceImpl.java
@@ -1,110 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright 2020 Crown Copyright (Health Education England)
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
- * associated documentation files (the "Software"), to deal in the Software without restriction,
- * including without limitation the rights to use, copy, modify, merge, publish, distribute,
- * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or
- * substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
- * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package uk.nhs.hee.trainee.details.service.impl;
-
-import java.util.Optional;
-import java.util.function.BiConsumer;
-import org.springframework.stereotype.Service;
-import uk.nhs.hee.trainee.details.mapper.TraineeProfileMapper;
-import uk.nhs.hee.trainee.details.model.PersonalDetails;
-import uk.nhs.hee.trainee.details.model.TraineeProfile;
-import uk.nhs.hee.trainee.details.service.PersonalDetailsService;
-import uk.nhs.hee.trainee.details.service.TraineeProfileService;
-
-@Service
-public class PersonalDetailsServiceImpl implements PersonalDetailsService {
-
-  private final TraineeProfileService service;
-  private TraineeProfileMapper mapper;
-
-  PersonalDetailsServiceImpl(TraineeProfileService service, TraineeProfileMapper mapper) {
-    this.service = service;
-    this.mapper = mapper;
-  }
-
-  @Override
-  public PersonalDetails createProfileOrUpdateBasicDetailsByTisId(String tisId,
-      PersonalDetails personalDetails) {
-    Optional<PersonalDetails> updatedDetails = updatePersonalDetailsByTisId(tisId, personalDetails,
-        mapper::updateBasicDetails);
-
-    if (updatedDetails.isEmpty()) {
-      TraineeProfile traineeProfile = new TraineeProfile();
-      traineeProfile.setId(tisId);
-      mapper.updateBasicDetails(traineeProfile, personalDetails);
-      TraineeProfile savedProfile = service.save(traineeProfile);
-      return savedProfile.getPersonalDetails();
-    }
-
-    return updatedDetails.get();
-  }
-
-  @Override
-  public Optional<PersonalDetails> updateGdcDetailsByTisId(String tisId,
-      PersonalDetails personalDetails) {
-    return updatePersonalDetailsByTisId(tisId, personalDetails, mapper::updateGdcDetails);
-  }
-
-  @Override
-  public Optional<PersonalDetails> updateGmcDetailsByTisId(String tisId,
-      PersonalDetails personalDetails) {
-    return updatePersonalDetailsByTisId(tisId, personalDetails, mapper::updateGmcDetails);
-  }
-
-  @Override
-  public Optional<PersonalDetails> updatePersonOwnerByTisId(String tisId,
-      PersonalDetails personalDetails) {
-    return updatePersonalDetailsByTisId(tisId, personalDetails, mapper::updatePersonOwner);
-  }
-
-  @Override
-  public Optional<PersonalDetails> updateContactDetailsByTisId(String tisId,
-      PersonalDetails personalDetails) {
-    return updatePersonalDetailsByTisId(tisId, personalDetails, mapper::updateContactDetails);
-  }
-
-  @Override
-  public Optional<PersonalDetails> updatePersonalInfoByTisId(String tisId,
-      PersonalDetails personalDetails) {
-    return updatePersonalDetailsByTisId(tisId, personalDetails, mapper::updatePersonalInfo);
-  }
-
-  /**
-   * Update the Personal Details entity for the given TIS ID.
-   *
-   * @param tisId           The TIS id of the trainee.
-   * @param personalDetails The personal details to add to the trainee.
-   * @param updateFunction  The function to use to update the personal details.
-   * @return The updated personal details or empty if a trainee with the ID was not found.
-   */
-  private Optional<PersonalDetails> updatePersonalDetailsByTisId(String tisId,
-      PersonalDetails personalDetails, BiConsumer<TraineeProfile, PersonalDetails> updateFunction) {
-    TraineeProfile traineeProfile = service.getTraineeProfileByTraineeTisId(tisId);
-
-    if (traineeProfile == null) {
-      return Optional.empty();
-    }
-
-    updateFunction.accept(traineeProfile, personalDetails);
-    return Optional.of(service.save(traineeProfile).getPersonalDetails());
-  }
-}
